quick fix
git-svn-id: http://data-race-test.googlecode.com/svn/trunk@2234 93a3164d-3a44-0410-8ae5-4baa038b0729

+++ tsan/ts_pin.cc
@@ -298,7 +298,7 @@ static void TLEBFlushUnlocked(ThreadLocalEventBuffer &tleb) {
 
   if (1 || DEBUG_MODE) {
     size_t max_idx = TS_ARRAY_SIZE(G_stats->tleb_flush);
-    size_t idx = min(u32_log2(tleb.size), max_idx - 1);
+    size_t idx = min((size_t)u32_log2(tleb.size), max_idx - 1);
     CHECK(idx < max_idx);
     G_stats->tleb_flush[idx]++;
   }
